string_get_size tests
This is a small collection of files to test various approaches to
implementing the string_get_size helper. test.c contains a driver
which excercises string_get_size() with a range of parameters - link
with your favorite implementation and run it. Pipe it through
check_get_size.py to see where it produces an inconsistent result.